Handle encoded u64 as `NUMERIC`
While SQLite cannot store u64 it can work with them, but only if they are cast as NUMERIC I think.

`SELECT CAST("18446744073709551610" as INTEGER) BETWEEN 18446744073709551609 AND 18446744073709551611;` => 0
`SELECT CAST("18446744073709551610" as NUMERIC) BETWEEN 18446744073709551609 AND 18446744073709551611;` => 1

+++ aquadoggo/src/db/stores/entry.rs
@@ -256,7 +256,7 @@ impl EntryStore<StorageEntry> for SqlStorage {
                 author = $1
                 AND log_id = $2
             ORDER BY
-                CAST(seq_num AS INTEGER) DESC
+                CAST(seq_num AS NUMERIC) DESC
             LIMIT
                 1
             ",
@@ -334,9 +334,9 @@ impl EntryStore<StorageEntry> for SqlStorage {
             WHERE
                 author = $1
                 AND log_id = $2
-                AND CAST(seq_num AS INTEGER) BETWEEN $3 and $4
+                AND CAST(seq_num AS NUMERIC) BETWEEN $3 and $4
             ORDER BY
-                CAST(seq_num AS INTEGER)
+                CAST(seq_num AS NUMERIC)
             ",
         )
         .bind(author.as_str())
@@ -387,9 +387,9 @@ impl EntryStore<StorageEntry> for SqlStorage {
             WHERE
                 author = $1
                 AND log_id = $2
-                AND CAST(seq_num AS INTEGER) IN ({})
+                AND CAST(seq_num AS NUMERIC) IN ({})
             ORDER BY
-                CAST(seq_num AS INTEGER) DESC
+                CAST(seq_num AS NUMERIC) DESC
             ",
             cert_pool_seq_nums
         );
